+++ module-huawei/src/main/java/com/affise/attribution/module/huawei/parameters/OaidMD5Provider.kt
@@ -1,8 +1,8 @@
-package com.affise.attribution.module.advertising.parameters
+package com.affise.attribution.module.huawei.parameters
 
 import com.affise.attribution.converter.Converter
 import com.affise.attribution.parameters.base.StringPropertyProvider
-import com.affise.attribution.module.advertising.oaid.OaidManager
+import com.affise.attribution.module.huawei.oaid.OaidManager
 import com.affise.attribution.parameters.ProviderType
 
 /**
@@ -13,10 +13,10 @@ import com.affise.attribution.parameters.ProviderType
  */
 class OaidMD5Provider(
     private val oaidManager: OaidManager,
-    private val converter: Converter<@JvmSuppressWildcards String, @JvmSuppressWildcards String>
+    private val converter: Converter<@JvmSuppressWildcards String, @JvmSuppressWildcards String>?
 ) : StringPropertyProvider() {
-    override val order: Float = 31.6f
+    override val order: Float = 31.61f
     override val key: ProviderType = ProviderType.OAID_MD5
 
-    override fun provide(): String? = oaidManager.getOaid()?.let(converter::convert)
+    override fun provide(): String? = oaidManager.getOaid()?.let { converter?.convert(it) }
 }