+++ module-huawei/src/main/java/com/affise/attribution/module/huawei/parameters/OaidProvider.kt
@@ -1,7 +1,7 @@
-package com.affise.attribution.module.advertising.parameters
+package com.affise.attribution.module.huawei.parameters
 
+import com.affise.attribution.module.huawei.oaid.OaidManager
 import com.affise.attribution.parameters.base.StringPropertyProvider
-import com.affise.attribution.module.advertising.oaid.OaidManager
 import com.affise.attribution.parameters.ProviderType
 
 /**
@@ -12,7 +12,7 @@ import com.affise.attribution.parameters.ProviderType
 class OaidProvider(
     private val oaidManager: OaidManager
 ) : StringPropertyProvider() {
-    override val order: Float = 31.7f
+    override val order: Float = 31.6f
     override val key: ProviderType = ProviderType.OAID
 
     override fun provide(): String? = oaidManager.getOaid()